+++ src/utils/operations/tests/capitalize.test.ts
@@ -1,7 +1,7 @@
 import { capitalize } from '../capitalize';
 
 describe('capitalize tests', () => {
-  const str = 'dharitrI';
+  const str = 'DharitrI';
 
   it('capitalizes the word correctly', () => {
     expect(capitalize(str)).toStrictEqual('DharitrI');